Add README.md and LICENSE files
Added README.md and LICENSE files to the root of the repository.

The README.md file will provides an overview of the gaming framework, including its purpose, features and setup instructions.

The LICENSE file will includes the licensing terms under which the project's code is distributed, ensuring clarity on usage rights and restrictions.